move doctor utility to utils/ directory
