+++ mlflow/__init__.py
@@ -89,7 +89,6 @@ xgboost = LazyLoader("mlflow.xgboost", globals(), "mlflow.xgboost")
 
 _configure_mlflow_loggers(root_module_name=__name__)
 
-from mlflow._doctor import doctor
 from mlflow.client import MlflowClient
 from mlflow.exceptions import MlflowException
 from mlflow.models import evaluate
@@ -152,6 +151,7 @@ from mlflow.tracking.fluent import (
 )
 from mlflow.utils.async_logging.run_operations import RunOperations  # noqa: F401
 from mlflow.utils.credentials import login
+from mlflow.utils.doctor import doctor
 
 __all__ = [
     "ActiveRun",
