+++ tests/utils/test_doctor.py
@@ -19,7 +19,7 @@ def test_doctor_active_run(capsys):
 def test_doctor_databricks_runtime(capsys):
     mock_version = "12.0-cpu-ml-scala2.12"
     with mock.patch(
-        "mlflow._doctor.get_databricks_runtime", return_value=mock_version
+        "mlflow.utils.doctor.get_databricks_runtime", return_value=mock_version
     ) as mock_runtime:
         mlflow.doctor()
         mock_runtime.assert_called_once()
